+++ CHANGELOG.md
@@ -14,6 +14,7 @@ under development
 - Add `Context.QueryString`.
 - Add `Context.QueryParams` and `Context.QueryParam`.
 - Add `Context.JSONP` and `Context.JSONPCallback`.
+- Add `Context.Render` to render a template.
 
 v1.8.1 April 2, 2020
 --------------------